Revert "Happychat: Refactor to use middleware"
