+++ client/lib/happychat/connection.js
@@ -17,20 +17,10 @@ class Connection extends EventEmitter {
 		if ( ! this.openSocket ) {
 			this.openSocket = new Promise( resolve => {
 				const url = config( 'happychat_url' );
-				debug( 'connecting to', url );
 				const socket = new IO( url );
 				socket
-					.on( 'init', () => {
-						resolve( socket );
-						debug( 'init', this );
-						this.emit( 'connect' );
-					} )
-					.on( 'connect', () => {
-						debug( 'connected' );
-						this.emit( 'connect' );
-					} )
-					.on( 'disconnect', () => this.emit( 'disconnect' ) )
-					.on( 'reconnect', () => this.emit( 'reconnect' ) )
+					.once( 'connect', () => debug( 'connected' ) )
+					.on( 'init', () => resolve( socket ) )
 					.on( 'token', handler => {
 						handler( { signer_user_id: user_id, jwt: token } );
 					} )
