+++ client/state/action-types.js
@@ -127,20 +127,16 @@ export const HAPPINESS_ENGINEERS_FETCH = 'HAPPINESS_ENGINEERS_FETCH';
 export const HAPPINESS_ENGINEERS_FETCH_FAILURE = 'HAPPINESS_ENGINEERS_FETCH_FAILURE';
 export const HAPPINESS_ENGINEERS_FETCH_SUCCESS = 'HAPPINESS_ENGINEERS_FETCH_SUCCESS';
 export const HAPPINESS_ENGINEERS_RECEIVE = 'HAPPINESS_ENGINEERS_RECEIVE';
-export const HAPPYCHAT_CONNECTION_OPEN = 'HAPPYCHAT_CONNECTION_OPEN';
 export const HAPPYCHAT_CONNECT = 'HAPPYCHAT_CONNECT';
 export const HAPPYCHAT_CONNECTED = 'HAPPYCHAT_CONNECTED';
 export const HAPPYCHAT_CONNECTING = 'HAPPYCHAT_CONNECTING';
-export const HAPPYCHAT_DISCONNECTED = 'HAPPYCHAT_DISCONNECTED';
 export const HAPPYCHAT_MINIMIZING = 'HAPPYCHAT_MINIMIZING';
 export const HAPPYCHAT_OPEN = 'HAPPYCHAT_OPEN';
 export const HAPPYCHAT_RECEIVE_EVENT = 'HAPPYCHAT_RECEIVE_EVENT';
-export const HAPPYCHAT_SEND_MESSAGE = 'HAPPYCHAT_SEND_MESSAGE';
+export const HAPPYCHAT_RECEIVE_TRANSCRIPT = 'HAPPYCHAT_RECEIVE_TRANSCRIPT';
 export const HAPPYCHAT_SET_AVAILABLE = 'HAPPYCHAT_SET_AVAILABLE';
 export const HAPPYCHAT_SET_CHAT_STATUS = 'HAPPYCHAT_SET_CHAT_STATUS';
 export const HAPPYCHAT_SET_MESSAGE = 'HAPPYCHAT_SET_MESSAGE';
-export const HAPPYCHAT_TRANSCRIPT_RECEIVE = 'HAPPYCHAT_TRANSCRIPT_RECEIVE';
-export const HAPPYCHAT_TRANSCRIPT_REQUEST = 'HAPPYCHAT_TRANSCRIPT_REQUEST';
 export const HELP_COURSES_RECEIVE = 'HELP_COURSES_RECEIVE';
 export const HELP_TICKET_CONFIGURATION_DISMISS_ERROR = 'HELP_TICKET_CONFIGURATION_DISMISS_ERROR';
 export const HELP_TICKET_CONFIGURATION_REQUEST = 'HELP_TICKET_CONFIGURATION_REQUEST';
